Adding the current Fedora sources
These are the latest sources used to build shim packages in Fedora.
shim-signed is from the f26 branch, shim and shim-unsigned-aarch64 are
one commit behind as shim includes a patch which was never built and
shim-unsigned-aarch64 was bumped for a mass rebuild that didn't pass.

We'll track changes from here.

Signed-off-by: Petr Šabata <contyk@redhat.com>